move safekey to identity mod
