+++ Cargo.toml
@@ -20,7 +20,7 @@ sha3 = "~0.8.2"
 threshold_crypto = "~0.3.2"
 tiny-keccak = "~1.5.0"
 unwrap = "~1.2.1"
-crdts = { git = "https://github.com/rust-crdt/rust-crdt", rev= "eab41f7" }
+crdts = "4.0.0"
 
   [dependencies.serde]
   version = "1.0.91"
