+++ src/identity/mod.rs
@@ -11,10 +11,56 @@ pub mod app;
 pub mod client;
 pub mod node;
 
-use crate::{utils, PublicKey, Result, XorName};
+use crate::{utils, AppFullId, ClientFullId, PublicKey, Result, Signature, XorName};
 use multibase::Decodable;
 use serde::{Deserialize, Serialize};
 use std::fmt::{self, Debug, Display, Formatter};
+use std::sync::Arc;
+
+/// An enum representing the Full Id variants for a Client or App.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum SafeKey {
+    /// Represents an application authorised by a client.
+    App(Arc<AppFullId>),
+    /// Represents a network client.
+    Client(Arc<ClientFullId>),
+}
+
+impl SafeKey {
+    /// Creates a client full ID.
+    pub fn client(full_id: ClientFullId) -> Self {
+        Self::Client(Arc::new(full_id))
+    }
+
+    /// Creates an app full ID.
+    pub fn app(full_id: AppFullId) -> Self {
+        Self::App(Arc::new(full_id))
+    }
+
+    /// Signs a given message using the App / Client full id as required.
+    pub fn sign(&self, msg: &[u8]) -> Signature {
+        match self {
+            Self::App(app_full_id) => app_full_id.sign(msg),
+            Self::Client(client_full_id) => client_full_id.sign(msg),
+        }
+    }
+
+    /// Returns a corresponding public ID.
+    pub fn public_id(&self) -> PublicId {
+        match self {
+            Self::App(app_full_id) => PublicId::App(app_full_id.public_id().clone()),
+            Self::Client(client_full_id) => PublicId::Client(client_full_id.public_id().clone()),
+        }
+    }
+
+    /// Returns a corresponding public key.
+    pub fn public_key(&self) -> PublicKey {
+        match self {
+            Self::App(app_full_id) => *app_full_id.public_id().public_key(),
+            Self::Client(client_full_id) => *client_full_id.public_id().public_key(),
+        }
+    }
+}
 
 /// An enum representing the identity of a network Node or Client.
 ///
